fix: fixed user specific data fetch service and target row bugs

+++ src/services/apiTasks.js
@@ -1,16 +1,25 @@
+import getUserId from "./getUserId";
 import supabase from "./supabase";
 
 export async function getTaskRemote() {
-  let { data: tasks, error } = await supabase.from("tasks").select("*");
+  const user_id = getUserId();
+  let { data: tasks, error } = await supabase
+    .from("tasks")
+    .select("*")
+    .eq("user_id", user_id);
   if (error) console.log(error);
   return tasks;
 }
 
 export async function addTaskRemote(newTask) {
+  const user_id = getUserId();
   if (Array.isArray(newTask)) {
+    const updatedNewTask = newTask.map((task) => {
+      return { ...task, user_id: user_id };
+    });
     const { data, error } = await supabase
       .from("tasks")
-      .insert(newTask)
+      .insert(updatedNewTask)
       .select();
     if (error) {
       console.error(error);
@@ -19,7 +28,7 @@ export async function addTaskRemote(newTask) {
   } else {
     const { data, error } = await supabase
       .from("tasks")
-      .insert([newTask])
+      .insert([{ ...newTask, user_id: user_id }])
       .select();
     if (error) {
       console.error(error);
@@ -29,19 +38,23 @@ export async function addTaskRemote(newTask) {
 }
 
 export async function updateTaskRemote(global_id, updatedTask) {
+  const user_id = getUserId();
   const { data, error } = await supabase
     .from("tasks")
     .update({ ...updatedTask })
     .eq("global_id", global_id)
+    .eq("user_id", user_id)
     .select();
   if (error) console.error(error);
   return data;
 }
 
 export async function deleteTaskRemote(global_id) {
+  const user_id = getUserId();
   const { error } = await supabase
     .from("tasks")
     .delete()
-    .eq("global_id", global_id);
+    .eq("global_id", global_id)
+    .eq("user_id", user_id);
   if (error) console.error(error);
 }
